test(backup): keep the simulated stuck isolate asleep on Linux
The kill-resistance tests park a worker in libc sleep() to stand in for a
native call that ignores Isolate.kill. On Linux the Dart VM profiler samples
threads with SIGPROF, and sleep() returns on any signal, so the "stuck"
isolate finished in a millisecond and five tests failed on CI (never seen
before because `dart analyze` used to fail first and skip `flutter test`).

Share one helper between the runner tests and RestoreBundleStaging's debug
hang that blocks the thread's signals before sleeping.
